+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
+### Added
+- "Launch at login" checkbox in the popover, backed by `SMAppService` (available when running from the built `.app` bundle).
 
 ## [1.0.0] - 2026-08-01
 ### Added